refactor: 목표의 메뉴 버튼 클릭시 반응 변경 등
- 목표의 메뉴 버튼을 누르면 드롭다운 다이얼로그가 뜨도록 변경
- 목표 추가인지 목표 수정인지 구분할 수 있도록 모드 번호 추가 및 전달함
- 목표 모델의 현재치와 목표치를 Int형으로 변경
- 테스트 코드 내용 보충

+++ app/src/main/java/com/minerdev/exermate/model/Goal.kt
@@ -9,8 +9,8 @@ data class Goal(
     var id: Int = 0,
     var createdAt: String = "",
     var updatedAt: String = "",
-    var current: Float = 0F,
-    var goal: Float = 0F,
+    var current: Int = 0,
+    var goal: Int = 0,
     var state: Byte = 0,
     var type: String = ""
 ) {
